RDMA/core: Inject Send Queue errors
Injecting an L_key from a foreign PD triggers a Local Protection
Error on a Send WR.

Because ib_post_send() has side-effects when returning an immediate
error, we are not at liberty to use the generic fail_function error
injection mechanism. When ib_post_send() fails with an immediate
error, it is supposed to set @bad_send_wr to the first WR that was
not posted.

Thus an immediate error is simulated by setting num_sge in one of
the posted WRs to an impossible value. The provider detects this and
throw an immediate error.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>