NFSD: Add tracepoints to record fattr4 results
Because fattr4 attributes are handled by setting bits in a mask
so that the server returns a list of individual attributes, I've
added a tracepoint to record the value of each attribute that
returns a non-constant (except for the more complex attribute
types such as fs_locations, which can be added later).

This is perhaps a noisy way to report attributes, but all of
these tracepoints are named so they can easily be disabled with
a glob.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>