Fix handling of POSIX draft default ACLs
The NFSv4.2 code does a SETATTR with a POSIX draft default ACL
of zero ACEs to delete the default ACL.  This patch fixes handling
of this case.

Note that the server code does not handle the case where
a file system is of ACL_SCOPE_FILE OBJECT (which means that
NFS4 and POSIX draft ACLs are mixed as the "true form" ACL
for different file objects in the file system).

Signed-off-by: Rick Macklem <rmacklem@uoguelph.ca>