lib: Add a "tagset" data structure
Introduce an ADT that acts as a set of strings. The typical use case
will be handling resource tagging (ie, metadata about a resource
that is represented simply by a name).

A tagset object today will consist solely of an xarray. However,
later we can add a Bloom or quotient filter to accelerate operations
such as "tagset_is_element" or "tagset_intersection".

A number of methods are introduced along with the new data type, but
consumers of the new APIs will be introduced in subsequent patches.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>