NFSD: Use xdrgen XDR functions for the NFSv3 LOOKUP procedure
Replace the NFSPROC3_LOOKUP entry in the nfs3_procedures array
with an entry that uses the xdrgen-built XDR functions. Introduce
a helper function to fill post_op_attr structures that will be
useful here and in subsequent patches.

The .pc_argzero field is now set to zero for NFSv3 LOOKUP
procedures. The xdrgen decoders are trusted to initialize all
arguments in the argp->xdr field, making the early defensive memset
unnecessary. The remaining argp fields are cleared as needed.

Because nfsd3_proc_lookup() now fills in the post_op_attrs and
stores the returned file handle in long-lived storage, the
fh_getattr() calls are done in the proc function and not in the XDR
result encoder. A separate ->pc_release callback is no longer
needed. This refactor makes it easy to confirm that the file handle
is released during every error flow.

Following the precedent of commit 40ee5dc6af35 ("knfsd: 64 bit ino
support for NFS server"), lease_get_mtime() determines the file's
mtime when filling in post-op attributes), but is not used for WCC
data. This seems odd to me; NFSv4 does not use lease_get_mtime() at
all, and not using it for WCC data seems like the file's reported
mtime values might vary depending on how the file metadata is
queried.

nfs3svc_encode_lookupres() no longer has any callers, so it is
removed.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>