NFSD: Use xdrgen XDR functions for NFSv3 ACCESS procedure
Replace the NFSPROC3_ACCESS entry in the nfs3_procedures array with
an entry that uses the xdrgen-built XDR functions.

The .pc_argzero field is now set to zero for NFSv3 ACCESS
procedures. The xdrgen decoders are trusted to initialize all
arguments in the argp->xdr field, making the early defensive memset
unnecessary. The remaining argp fields are cleared as needed.

Because nfsd3_proc_access() now fills in the post_op_attrs, the
fh_getattr() calls are done in the proc function and not in the XDR
result encoder. A separate ->pc_release callback is no longer
needed. This refactor makes it easy to confirm that the file handle
is released during all error flows.

nfs3svc_decode_accessargs() and nfs3svc_encode_accessres() no longer
have any callers, so they are removed.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>