NFSD: Use xdrgen XDR functions for NFSv3 GETATTR procedure
Replace the NFSPROC3_GETATTR entry in the nfs3_procedures array with
an entry that uses the xdrgen-built XDR functions. A few helper
functions are introduced which will be used here and in subsequent
patches.

The .pc_argzero field is now set to zero for NFSv3 GETATTR
procedures. The xdrgen decoders are trusted to initialize all
arguments in the argp->xdr field, making the early defensive memset
unnecessary. The remaining argp fields are cleared as needed.

Because nfsd3_proc_getattr() now fills in the response attributes,
the fh_getattr() calls are done in the proc function and not in the
XDR result encoder. A separate ->pc_release callback is no longer
needed. This refactor makes it easy to confirm that the file handle
is released during every error flow.

Following the precedent of commit 40ee5dc6af35 ("knfsd: 64 bit ino
support for NFS server"), lease_get_mtime() determines the file's
mtime for GETATTR (and later when filling in post-op attributes),
but is not used for WCC data. This seems odd to me; NFSv4 does not
use lease_get_mtime() at all, and not using it for WCC data seems
like an NFSv3 file's reported mtime values might vary depending on
how the file metadata is queried.

nfs3svc_encode_getattrres() no longer has any callers, so it is
removed.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>