NFSD: Use xdrgen XDR functions for NFSv3 SETATTR procedure
Replace the NFS3PROC_SETATTR entry in the nfs3_procedures array with
an entry that uses the xdrgen-built XDR functions. Introduce a
helper function to convert from struct sattr3 to struct iattr.

The .pc_argzero field is now set to zero for NFSv3 SETATTR
procedures. The xdrgen decoders are trusted to initialize all
arguments in the argp->xdr field, making the early defensive memset
unnecessary. The remaining argp fields are cleared as needed.

Because nfsd3_proc_setattr() now fills in the WCC metadata, the
fh_getattr() calls are done in the proc function and not in the XDR
result encoder. A separate ->pc_release callback is no longer
needed. This refactor makes it easy to confirm that the file handle
is released during every error flow.

nfs3svc_decode_sattrargs() no longer has any callers, so it is
removed.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>