NFSD: Replace nfs3.h with nfs3xdr_gen.h
nfs3xdr_gen.h includes a set of header files that are tool-generated
from the nfs3.x specification. The goal is to replace NFSD's XDR
infrastructure with tool-generated files which are to have better
memory safety and are less brittle to maintain.

Rename or move some symbolic constants to enable NFSD to continue
building properly.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>