NFSD: Use xdrgen XDR functions for NFSv3 WRITE procedure
Replace the NFSPROC3_WRITE entry in the nfs3_procedures array
with an entry that uses the xdrgen-built XDR functions.

Special handling is needed to decode the write payload, which
will be directed into the server's page cache.

Clamping the payload size when decoding serves no functional purpose
thus it is removed entirely. The transport layer has already
guaranteed that the incoming message (including the WRITE payload)
does not exceed svc_max_payload.

The WRITE count versus data length sanity check is moved to the proc
function: if those values are ever not equal, that does not count as
GARBAGE_ARGS, since the server /is/ able to decode these arguments.
Rather the arguments are invalid.

The .pc_argzero field is now set to zero for NFSv3 WRITE
procedures. The xdrgen decoders are trusted to initialize all
arguments in the argp->xdr field, making the early defensive memset
unnecessary. The remaining argp fields are cleared as needed.

Because nfsd3_proc_write() now fills in the WCC metadata,
including both pre_op_attr and post_op_attr, the fh_getattr()
and fh_pre_saved operations are handled in the proc function
and not in the XDR result encoder. A separate ->pc_release
callback is no longer needed. This refactor makes it straightforward
to confirm that the file handle is released during all error flows.

nfs3svc_decode_writeargs() and nfs3svc_encode_writeres no longer
have any callers, so they are removed.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>