NFSD: Use xdrgen XDR functions for NFSv3 READLINK procedure
Replace the NFSPROC3_READLINK entry in the nfs3_procedures array
with an entry that uses the xdrgen-built XDR functions.

Special handling is needed to XDR encode the link's content, which
resides in the server's page cache, while avoiding a data copy.

The .pc_argzero field is now set to zero for NFSv3 READLINK
procedures. The xdrgen decoders are trusted to initialize all
arguments in the argp->xdr field, making the early defensive memset
unnecessary. The remaining argp fields are cleared as needed.

Because nfsd3_proc_readlink() now fills in the post_op_attrs, the
fh_getattr() calls are done in the proc function and not in the XDR
result encoder. A separate ->pc_release callback is no longer
needed. This refactor makes it easy to confirm that the file handle
is released during all error flows.

nfs3svc_encode_readlinkres() no longer has any callers, so it is
removed.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>